macos: pin legacy objc deps to version before mass deprecation
These crates are all deprecated in favor of the new objc2 crate,
but the migration to objc2 is non-trivial.  Meanwhile, the volume
of deprecation warnings is overwhelming and getting in the way.

Pin these deps back to a versions prior to the mass deprecation
so that we're not drowning in spew in the meantime.